kbc-quiz-app-simplified-code/requirements.txt
customtkinter==5.2.0
Pillow==9.5.0 
pygame==2.5.0
darkdetect==0.8.0
